feat: correção de codigo de acordo com solicitação no PR
