+++ src/services/UserService.ts
@@ -1,10 +1,8 @@
 import jwt from 'jsonwebtoken';
 import bcrypt from 'bcrypt';
-import { hash } from 'bcrypt';
 import { Repository } from 'typeorm';
 import { MysqlDataSource } from '../config/database';
 import { User } from '../database/entity/User';
-import { randomInt } from 'crypto';
 
 export class UserService {
   private userRepository: Repository<User>;
@@ -45,15 +43,17 @@ export class UserService {
     return await this.userRepository.findOne({ where: { id } });
   }
 
-  async updatePassword(id: number, password: string): Promise<string> {
+  async updatePassword(
+    id: number,
+    password: string
+  ): Promise<User | undefined> {
     const user = await this.userRepository.findOne({ where: { id } });
     if (!user) {
       return undefined;
     }
 
     if (user) {
-      const randomSalt = randomInt(10, 16);
-      const passwordHash = await hash(password, randomSalt);
+      const passwordHash = await bcrypt.hash(password, 10);
       user.password = passwordHash;
       this.userRepository.save(user);
     }
